Move Kindergeldübertrag functions to new module.
